Unsigned integer arithmetic

+++ runtime-jvm/main/src/main/scala/Term.scala
@@ -444,8 +444,8 @@ object Term {
     }
 
     implicit def bool(b: Boolean): Term = Unboxed(boolToUnboxed(b), UnboxedType.Boolean)
-    implicit def number(n: Long): Term = Unboxed(longToUnboxed(n), UnboxedType.Integer)
-    implicit def number(n: Int): Term = Unboxed(intToUnboxed(n), UnboxedType.Integer)
+    implicit def number(n: Long): Term = Unboxed(longToUnboxed(n), UnboxedType.Int64)
+    implicit def number(n: Int): Term = Unboxed(intToUnboxed(n), UnboxedType.Int64)
     implicit def double(n: Double): Term = Unboxed(doubleToUnboxed(n), UnboxedType.Float)
     implicit def stringAsText(s: String): Term = Text(util.Text.fromString(s))
     implicit def nameAsVar(s: Name): Term = Var(s)
